coding_standard.md: Relax warning as error on release (opensource)
This is a "user oriented" best practice used in many FLOSS projects.
Fixing at least one supported env is good, but maybe not enough,
extending support can help but will it scale?

Reality of sofware engineering is that software integration
is a moving science, constraints are evolving over space and time.
We can't anticipate (future) breakage,
this change can mitigate the annoyance and maximize user adoption,
help integrators and ease software maintenance.

The alternative option, would be to wait user(s) to complain
and provide support for the desired env in an acceptable time frame.

Signed-off-by: Philippe Coval <philippe.coval@silabs.com>